перед сборкой
