+++ components/FormValidator.js
@@ -1,16 +1,16 @@
 export default class FormValidator {
-  constructor(config, form) {
+  constructor(settings, form) {
     this._form = form;
-    this._inputSelector = config.inputSelector;
-    this._submitButtonSelector = config.submitButtonSelector;
-    this._inactiveButtonClass = config.inactiveButtonClass;
-    this._inputErrorClass = config.inputErrorClass;
-    this._errorClass = config.errorClass;
+    this._inputSelector = settings.inputSelector;
+    this._submitButtonSelector = settings.submitButtonSelector;
+    this._inactiveButtonClass = settings.inactiveButtonClass;
+    this._inputErrorClass = settings.inputErrorClass;
+    this._errorClass = settings.errorClass;
     this._buttonElement = form.querySelector(this._submitButtonSelector);
     this._inputList = Array.from(this._form.querySelectorAll(this._inputSelector));
   }
 
-/*показывает элемент ошибки!!!*/
+/*показывает элемент ошибки*/
   _showInputError(inputElement) {
     const errorElement = this._form.querySelector(`.${inputElement.id}-error`);
     inputElement.classList.add(this._inputErrorClass);
@@ -18,14 +18,14 @@ export default class FormValidator {
     errorElement.classList.add(this._errorClass);
   }
 
-/*скрывает элемент ошибки!!!*/
+/*скрывает элемент ошибки*/
 _hideInputError(inputElement) {
   const errorElement = this._form.querySelector(`.${inputElement.id}-error`);
   inputElement.classList.remove(this._inputErrorClass);
   errorElement.classList.remove(this._errorClass);
   errorElement.textContent = '';
 }
-/*Обработчик форм!!! */
+
 _setEventListners() {
   this.toggleButtonState();
   this._inputList.forEach((inputElement) => {
@@ -36,7 +36,7 @@ _setEventListners() {
   });
 }
 
-/* блокировка кнопки!!!*/
+/* блокировка кнопки*/
 toggleButtonState() {
   this._buttonElement.disabled = !this._form.checkValidity();
   if (this._buttonElement.disabled) {
@@ -46,7 +46,7 @@ toggleButtonState() {
     this._buttonElement.classList.remove(this._inactiveButtonClass);
   }
 }
-/*проверка на валидность!!!*/
+
 _handleField(inputElement) {
   if (!inputElement.validity.valid) {
     this._showInputError(inputElement);
@@ -55,22 +55,10 @@ _handleField(inputElement) {
   }
 }
 
-  _disabledButton () {
-    this._buttonElement.classList.add(this._config.inactiveButtonClass);
-    this._buttonElement.setAttribute('disabled', 'disabled');
-  };
-
-  _hasInvalidInput () {
-    return this._inputList.some((inputElement) => {
-    return !inputElement.validity.valid;
-  });
-  };
   enableValidation() {
-    this._form.addEventListener('submit', () => {
-      this._disabledButton();
-    });    
-    this._setEventListners();    
-  };
+    this._setEventListners(this._form);
+  }
+
   resetValidation() {
     this.toggleButtonState();
 
@@ -78,5 +66,4 @@ _handleField(inputElement) {
      this._hideInputError(inputElement)
     });
   }
-
 }
