+++ index.html
@@ -24,16 +24,16 @@
       </section>
       <section class="elements" aria-label="Фотогалерея"></section>
     </main>
-    <footer class="footer">&copy; 2022 Mesto Russia</footer>
+    <footer class="footer">&copy; 2023 Mesto Russia</footer>
   </div>
   <div class="popup popup_edit">
     <div class="popup__container">
       <button class="popup__close-button" type="button"></button>
       <h2 class="popup__header">Редактировать профиль</h2>
       <form name="profile-information" class="popup__form" novalidate>
-        <input type="text" class="popup__field popup__field_el_name" id="name-input" required placeholder="Имя" minlength="2" maxlength="40">
+        <input type="text" class="popup__field popup__field_el_name" id="name-input" name="name" required placeholder="Имя" minlength="2" maxlength="40">
         <span class='popup__form-error name-input-error'></span>
-        <input type="text" class="popup__field popup__field_el_job" id="job-input" required placeholder="О себе" minlength="2" maxlength="200">
+        <input type="text" class="popup__field popup__field_el_job" id="job-input" name="description" required placeholder="О себе" minlength="2" maxlength="200">
         <span class='popup__form-error job-input-error'></span>
         <button class="popup__submit-button" type="submit">Сохранить</button>
       </form>
@@ -44,9 +44,9 @@
       <button class="popup__close-button" type="button"></button>
       <h2 class="popup__header">Новое место</h2>
       <form name="profile-information" class="popup__form" novalidate>
-        <input type="text" class="popup__field popup__field_el_name" id="title-input" required placeholder="Название" minlength="2" maxlength="30">
+        <input type="text" class="popup__field popup__field_el_name" id="title-input" name="title" required placeholder="Название" minlength="2" maxlength="30">
         <span class='popup__form-error title-input-error'></span>
-        <input type="url" class="popup__field popup__field_el_link" id="url-input" required placeholder="Ссылка на картинку">
+        <input type="url" class="popup__field popup__field_el_link" id="url-input" name="url" required placeholder="Ссылка на картинку">
         <span class='popup__form-error url-input-error'></span>
         <button class="popup__submit-button" type="submit">Создать</button>
       </form>
@@ -69,6 +69,6 @@
       </div>
     </div>
   </template>
-  <script src="./scripts/index.js" type="module"></script>
+  <script src="./pages/index.js" type="module"></script>
 </body>
 </html>
